test(toggle): migrate color
