+++ core/src/components/toggle/test/color/toggle.e2e.ts-snapshots/toggle-color-checked-ios-ltr-Mobile-Firefox-linux.png
