+++ core/src/components/toggle/test/color/toggle.e2e.ts-snapshots/toggle-color-checked-ios-ltr-Mobile-Safari-linux.png
