+++ core/src/components/toggle/test/color/toggle.e2e.ts-snapshots/toggle-color-unchecked-ios-ltr-Mobile-Chrome-linux.png
