+++ core/src/components/toggle/test/color/toggle.e2e.ts-snapshots/toggle-color-unchecked-md-ltr-Mobile-Safari-linux.png
